Extract atomic file module
