+++ linux/user-daemon/src/stores/file_store_v2.rs
@@ -1,15 +1,13 @@
-use std::fs;
-use std::fs::{File, OpenOptions};
+use std::fs::File;
 use std::io;
 use std::io::Write;
-use std::os::unix::fs::OpenOptionsExt;
-use std::path::Path;
 use std::path::PathBuf;
 
 use directories::ProjectDirs;
 use serde_json;
 use u2f_core::{AppId, ApplicationKey, Counter, KeyHandle, SecretStore};
 
+use atomic_file;
 use stores::Secret;
 
 #[derive(Serialize, Deserialize)]
@@ -60,7 +58,7 @@ impl FileStoreV2 {
     }
 
     fn write(&self, data: &Data) -> io::Result<()> {
-        overwrite_file_atomic(&self.path, move |writer| {
+        atomic_file::overwrite(&self.path, move |writer| {
             serde_json::to_writer_pretty(writer, &data).map_err(|e| e.into())
         })
     }
@@ -101,51 +99,6 @@ impl SecretStore for FileStoreV2 {
             .find_secret(application, handle)
             .map(|secret| secret.application_key.clone()))
     }
-}
-
-fn overwrite_file_atomic<W>(path: &Path, writer_fn: W) -> io::Result<()>
-where
-    W: FnOnce(Box<&mut dyn Write>) -> io::Result<()>,
-{
-    let directory = path.parent().ok_or(io::Error::new(
-        io::ErrorKind::InvalidInput,
-        "Invalid file path, does not have a parent directory",
-    ))?;
-    let tmp_path = make_tmp_path(path)?;
-
-    {
-        let mut tmp_file = OpenOptions::new()
-            .write(true)
-            .create_new(true)
-            .mode(0o600)
-            .open(&tmp_path)?;
-        writer_fn(Box::new(&mut tmp_file))?;
-        tmp_file.flush()?;
-        tmp_file.sync_all()?;
-    }
-
-    fs::rename(&tmp_path, path)?;
-    fsync_dir(directory)?;
-    Ok(())
-}
-
-fn fsync_dir(dir: &Path) -> io::Result<()> {
-    let f = File::open(dir)?;
-    f.sync_all()
-}
-
-fn make_tmp_path(path: &Path) -> io::Result<PathBuf> {
-    let mut tmp_path = PathBuf::from(path);
-    let mut file_name = tmp_path
-        .file_name()
-        .ok_or(io::Error::new(
-            io::ErrorKind::InvalidInput,
-            "Invalid file path, does not end in a file name",
-        ))?
-        .to_owned();
-    file_name.push(".tmp");
-    tmp_path.set_file_name(file_name);
-    Ok(tmp_path)
 }
 
 #[cfg(test)]
